refactor(rspack_core): move compilation make to a struct
